finally sync 0o0
