+++ .env
@@ -1,0 +1,2 @@
+NEXT_PUBLIC_COUNTRY_CODE="UKR"
+NEXT_PUBLIC_CATEGORIES="153, 154"
