+++ .env.example
@@ -1,0 +1,2 @@
+NEXT_PUBLIC_COUNTRY_CODE=
+NEXT_PUBLIC_CATEGORIES=
