+++ package.json
@@ -17,6 +17,7 @@
     "@types/react-dom": "18.2.5",
     "autoprefixer": "10.4.14",
     "axios": "^1.4.0",
+    "classnames": "^2.3.2",
     "eslint": "8.43.0",
     "eslint-config-next": "13.4.6",
     "next": "13.4.6",
