+++ src/app/layout.tsx
@@ -1,8 +1,10 @@
+import Footer from "./(components)/Footer";
+import Header from "./(components)/Header";
 import "./globals.css";
-import { Inter } from "next/font/google";
+import { Montserrat } from "next/font/google";
 import ReactQueryWrapper from "@/utils/queryClientProvider";
 
-const inter = Inter({ subsets: ["latin"] });
+const montserrat = Montserrat({ subsets: ["latin"] });
 
 export const metadata = {
   title: "Create Next App",
@@ -16,8 +18,10 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body className={inter.className}>
+      <body className={montserrat.className}>
+        <Header />
         <ReactQueryWrapper>{children}</ReactQueryWrapper>
+        <Footer />
       </body>
     </html>
   );
