+++ src/utils/static-categories.ts
@@ -1,0 +1,3 @@
+export const ALL_CATEGORIES = process.env
+  .NEXT_PUBLIC_CATEGORIES!.split(",")
+  .map((cat) => parseFloat(cat));
